refactor(events): split TierForm under the line cap (#544)
Extract three cohesive presentational sub-components and one pure-helper
module from the 1133-line ticket-tier form — the campaign's largest
component. The full form state machine (state, effects, deriveds,
venues query, mutations, handleSubmit/handleDelete) stays byte-for-byte
in the parent; sub-sections receive parent-owned $state via $bindable +
bind: (the settled Task 4 pattern) and read-only data as plain props.

New files:
- tier-form-helpers.ts: CURRENCY_SYMBOLS, SUPPORTED_CURRENCIES,
  toDatetimeLocal, toTimezoneAwareISO, normalizeDecimalInput (verbatim).
- TierFormPricingSection.svelte: price type / currency / fixed / PWYC
  inputs (bindable priceType, currency, price, pwycMin, pwycMax).
- TierFormAvailabilitySection.svelte: quantity, sales period, visibility,
  purchasable-by, membership restriction (8 bindables + readbacks).
- TierFormSeatingSection.svelte: seat-assignment mode, max-tickets,
  venue readback + sector picker (bindable seatAssignmentMode,
  maxTicketsPerUser, sectorId; selectedVenue/sectors passed by value).

The #535 finalPrice branch control flow, the tier.id ?? tier.name each
keys, and the sectorValid submit gate are preserved unchanged. Zero
behavior change — extraction only.

Parent 1133 -> 597 lines (under the 750 cap, ~20% headroom).
